Hi-Five_Recognition/ml/requirements.txt
# Core ML
mediapipe>=0.10.14
xgboost>=2.0.0
scikit-learn>=1.3.0
numpy>=1.24.0
pandas>=2.0.0
joblib>=1.3.0

# Image / video I/O
opencv-python>=4.8.0
Pillow>=10.0.0

# Web server
Flask>=3.0.0
flask-cors>=4.0.0

# CLI niceties
tqdm>=4.65.0
